Fetch current timer each time in thinker loop.
This fixed a bug with timer appearing to go in reverse in some cases.  Most of the
time we pass around the current timestamp "usecNow" to avoid 1.) fetching the timestamp
again, which can be slow and 2) complicated multithreaded race conditions if we put it
in a global variable.  But in at least one place, when we ask a thinker to think,
it might pass into code that doesn't follow this convention and fetches the timestamp.
In particular, the fake lag thinker was calling the generic packet receive code, which
fetches the timestamp to record when the packet was received.  Then the next thinker would
run, using a timestamp that was earlier than the time when the packet was marked as
received.  This led to some asserts.  (And possibly worse).