Add symmetric cipher context to Crypto API.
This just wraps whatever context is needed by the crypto provider.  It should allow me to avoid paying the setup costs for each packet.  (Issue #47)

This should not only be a nice (small) optimization, it wil make it easier to support bcrypt (Issue #45).

We should update the perf tests to do encrypt/decrypt a bunch of buffers using the same cipher parameters, using these objects, since this models our use case in this networking library.

Next step is to use this in the CSteamNetworkConnectionBase, instead of storing raw keys.