Some refactoring to make way for the "Messages" API.
Steam has a higher level UDP-style API, currently called ISteamNetworkingMessages.  There are no connection handles, instead you just send a message to a peer.  (It still supports reliable messages and messages longer than datagrams, the difference is in how things are addressed at the API layer.)

The Messages interface creates connections internally, and those connections are not visible to the app through the normal connections interface.  So receiving messages and handling connection state change notifications needs to be dispatched differently.

This change introduces hooks in CSteamNetworkConnectionBase to better facilitate this.  It happened because, on the Steam side, all such internally created connections used to be a specific derived class.  Those hooks could then be accomplished by overriding virtual functions.  But I'm changing that to make it so that different connections might be used.  For example, you might use the Messages interface with IP address as the peer identity, which would make porting existing UDP much easier.  Or if you open up a session to send messages to yourself (or another local identity, say the gameserver when you are the user), we would use a pipe connection.  In this situation, basically any connection class could be used by the Messages layer, and using virtual functions as the means of hooking in won't work; hence these changes.  (If we were using a dynamic language like Python or Javascript, we could do it, though.)