Fix some minor bugs in cert store.
These bugs were discovered using the novel method of writing a test suite
and actually running it.  (Checking that in shortly.)

Also tweaked the #define that controls the single hardcoded root CA,
if any.  I do not have any test code for when we do not want to do this
right now.  For the open source code probably we want to leave it dynamic
by default.