Fix CUtlHashMap initialization bug.
Fix a pretty sharp edge: when inserting into CUtlHashMap with a function that
only takes a key and inserts the default value, we were using "default
initialization" ("new T"), which does not initialize PODs.  Switch to
"value initialization" ("new T{}").

Also some minor compiler compatibility fixes.  The version of gcc we're using
here didn't like the {}-style call of the base class constructor, so use ()
instead.  *shrug*

Also deleted a kludge for an older version of Visual Studio.  We were
supporting that for a while in Steam, but we've upgraded our min specs and
so I'm doing the same here.