Bump protocol version
This was needed for some changes to relayed connections.  There are no relevant
changes in the opensource code, but I want to keep thebranches in sync and this
is safe to do.  Old clients should never reject newer clients!  Instead, newer
clients should offer back compat to old clients, when possible.