Add k_nThinkTime_ASAP
Let's not use 0 for this anywhere, because we assert in the thinker system
if we try to set a think time to zero, since this is often an unitialized
variable.  Instead, we use the next lowest number, 1.  But let's not have
a hardcoded 1 anywhere, either, let's make it clear what the meaning of
the 1 is.