Fix bug not scheduling reliable ack.
The protocol still worked, as the sender would retry.  But this was not
optimal, and the sender was interpreting the lack of an ack as a problem,
which led to some confusing spew and stats not being quite right.

This was not detected earlier, because most of the use cases used for
testing are "game-like", in that they send packets nearly constantly.
However, I noticed right at the connection start, when the client is
loading, the server would consistently spew about a "timeout" happening
on a perfect connection.  This is because when the client is loading
the level, the connection is idle (and we rely on the protocol to do
keepalives, etc).