Track the "wire state" of a connection.
This is a subset of the connection state that does not go thorugh certain
transitions.  Having this simplifies certain logic, and I used it to fix
several edge cases where we were not cleaning up properly.

And example is CConnectionTransportUDPBase::SendConnectionClosedOrNoConnection,
but there were many more such places in the relay code.