Refactor accounting of sequenced packet stats.
For the billionth (and probably not the last) time.

Move the actual counters into a struct, so that we can count them at multiple
levels.  For example, on the relay, I want to easily aggregate these stats
by data center.

Also moved the accounting functions into the header and broke them up into
multiple functions, so different stats tracker types can "override" the
methods.  ("Override" in quotes here because we are doing this with templates
so everything is totally inline, with no virtual function dispatch, so the
compiler can optimize everytyhing.)