Refactor auth/cert code
Promoted a bunch of code to the base class, which previously in a derived
class not part of this opensource code.

Added a new define STEAMNETWORKINGSOCKETS_CAN_REQUEST_CERT, to make it
clear why some code is #defined in/out.  (Also, in some environments
where this code is used, we might not be able to *request* a cert, but
we can validate it and we have authentication mechanisms.)

This code also contains logic to refresh a cert if it's near expiry.
It's not active in the opensource code, since the opensource code doesn't
have a built in cert authority.  Once the code ships in the Steamworks
SDK, it will resolve #129, which was the entire purpose of doing all this
work in the first place.