Add trivial signaling client example.
WIP.  This compiles but hasn't been tested yet.  Working on a test case
that exercises it.

Also I'm still not sure I like the way the interfaces are organized.  It's
slightly more complicated than necessary, to handle what is probably a
relatively rare use case.  Maybe the only use case I care about is "Here
is the one and only signaling service to use" and I could essentially
combine ISteamNetworkingCustomSignalingRecvContext and
ISteamNetworkingCustomSignalingService.  You could replace the default
platform service, but not mix and match.
Or maybe the only thing that is needed is some comments clarifying
the relationship between ISteamNetworkingCustomSignalingRecvContext
and ISteamNetworkingCustomSignalingService.  In both cases, the primary
purpose is to create a signaling session for a specific connection.  But
one is for remotely initiated connections (a signal that respresents a
connect request coming in) and the other is for locally-initated
requests that do not go through ConnectP2PCustomSignaling.