P2P refactor for hosted dedicated server connections
This change has a bunch of refactoring to allow connections to be made on
Steam to servers in SDR known data centers.  Previously, the app needed to
have a central matchmaking authority (what we call the "game coordinator")
issue a ticket.  Now, the client may initiate the connection as an ordinary
P2P connect, not knowing that the server is in a known data center.  The
server will issue a ticket and sign it with its own cert and then send back
the ticket via signaling.  Then the client will switch to use the optimized
routing for this case.