Simple optimization to simulated packet lag
When finding the proper place to insert into the linked list, we were using
a dumb linear search -- which would be fine since in the steady state all
insertions are at one end and we don't need to search.  Only problem was
we were doing our search starting at the wrong end, so basically every
insertion would traverse the entire queue.