Use timeBeginPeriod on Windows
We need a high resolution timer so we can wake up at the right time when we
are pacing out packets, etc.  Without this call, we get the crappy default
resolution.  Up to this point it looks like we have just been getting saved
by some other code in the game, or Steam client, calling this for us.  But
really we need to do it ourselves.

https://randomascii.wordpress.com/2020/10/04/windows-timer-resolution-the-great-rule-change/

P4:6571383