Sync up with Steamworks SDK
With the exception of steam_api_common.h, which is mostly a stub,
these files are now identical with the Steamworks SDK.

Note that this also removes a check for STEAMNETWORKINGSOCKETS_OPENSOURCE
to try to set the structure packing consistently.  However, due to an
order-of-include bug (see issue #176), this wasn't actually working anyway,
and probably isn't the best approach.

P4:6610285