Initialize table of configuration values in CSteamNetworkingUtils constructor.
Don't try to do it lazily at any other time.
Added some asserts and clarified comments.

This fixes a bug if the first call you make is IterateGenericEditableConfigValues.
That could have been solved with a narrower fix, but I realized the lazy initialization
was really just not useful and doing it this way makes more sense.

P4:6854435,6854450