Bump protocol version for multi-lane support
Fail connection if we try to configure multiple lanes but peer is too old.

We might consider failing only if they actually sent a message on a
lane > 0, but this is simpler and I don't think in pradctice there will
be any meaningful difference to app code.  Failing immediately is probably
simplest.

p4:68744623