Improved vcpkg support
Moved instructions for building windows manually to a separate file
and marked that method as not supported.

- Deleted the vcpkg CONTROL file, upgrade to a modern vcpkg.json manifest.
- Added tests and examples as vcpkg "features".
- Make sure we obey static/shared from the vcpkg triplet (I believe this
  addresses #175)
- Split out the detailed instructions for how to build Windows dependencies
  manually to a separate document, so we don't clutter up the main doc.

Also:
- Examples will always link with shared lib
- Tests always link with the static lib

I'm doing my best here to obey vcpkg (and cmake in general) best
practices / conventions, but this is not really something I have
expertise in, and the documentation on all this is realy not good,
so feel free to make suggestions if you see something that could be
done better.

Possible future improvements (HELP WANTED):
- Don't select the crypto with a feature, that is explicitly listed
  in the vcpkg docs as an anti-pattern.
- Add a feature for the cert tool.
- Add more CI actions to build more platforms to make sure they
  are all working.