More WebRTC / ICE refactor
Move the generation of the ICESessionConfig into the common P2P code, since
none of it was specific to WebRTC.

Still need to do some refactoring of the Valve client, but checking in this
checkpoint as the code is working again.

Also, I'm aware that I have lost the configuration of the TURN servers.
Will add that back after I'm done with this big merge and refactor.  I hit
some merge conflicts.

P4:7194237