Fix a bug which would trigger under very high packet loss.
The basic problem was that we were using m_statsEndToEnd.m_nMaxRecvPktNum
for two purposes:
- The highest numbered we have received, for purposes of jitter calculations
  and duplicate packet rejection.
- The highest numbered packet we want to ack.

But if we decide to drop a packet (to protect against a malicious sender, etc)
and intentionally not ack it, these values won't be the same!  This means we
might be acking packets that we did not actually process!  Fortunately, it only
happened if the fragmentation was extremely high, which only happens when there
is extrenely high packet loss.  Also, only certain cases of "don't ack this"
can cause any problems other than asserts.

The "soak" connection test has been detecting this bug for a while, but I
haven't had a chance to investigate, since it only triggered underly insanely
high packet loss (whichthe test simulates).  But now the soak test passes,
which....is good!

With this change, the role of m_statsEndToEnd.m_nMaxRecvPktNum is only for
duplicate detection.  We track the highest numbered packet that we wish to ack
in the sentinel in m_receiverState.m_mapPacketGaps.  The time when that packet
was received will also be tracked there.  This change actually simplified some
code.

One more change: how we protect against a malicious sender causing our packet gap
map to be too large.  When we go to add a new gap, previously if we were already
at the max, we would immediately abort the function.  Now, we move the handling
of "too many gaps" to be later in the function, and we try to select the "best"
gap to erase, with the goal of minimizing the degregation (very possibly none)
caused by us starting to nack a block of packets that were actually received
(and in fact perhaps previousally acked).

P4:7219136