Replace several instance of CUtlVector with std::vector
You cannot put a std::string in a CUtlVector, because on some implementations
of STL, std::string does reference counting and other faqncy stuff, and cannot
be trivially relocated.  And I really want to get rid of this class in this
opensource library, to keep the dependencies and code size small and just use
more standard stuff.

P4:8088402,8088766,8088829