Refactor packet lagging.
This is in preparation to support automatically correcting out-of-order
packets.

Each packet in the list will now be dynamically allocated, and we will
use a hand coded linked list, instead of CUtlLinkedList.

Also, the lagged packet structure will be derived from
CPossibleOutOfOrderPacket.  If we only care about fake lag, then this
change is basically a net negative, as it adds dynamic memory overhead.
However, in real end user use cases, the only packets that will be
"lagged" are going to be those that we are holding in case they are
out of order.

P4:8760651